+++ packages/core/CHANGELOG.md
@@ -5,6 +5,8 @@
 - Schema validation added to `config.json`
 - Validation support added to create and update api
 - Search API 
+- New config added `"schema.validation": "strict" | "optional"`
 
 ### Changed
 - Updated `dev` setup to run example
+- server port in `config.json` is changed to lowercase `port`